Extract logic form issuer method in one class
ISSUE: MOCRSET04-9
