+++ GXModules/Mollie/Mollie/Shop/Themes/Honeygrid/checkout_payment_modules.html
@@ -1,7 +1,7 @@
 {block name="checkout_payment_modules_list"}
     <script src="/GXModules/Mollie/Mollie/Shop/Javascripts/mollie-applepay.js"></script>
-    <script src="/GXModules/Mollie/Mollie/Shop/Javascripts/mollie-credit-card.js"></script>
     <script src="/GXModules/Mollie/Mollie/Shop/Javascripts/mollie-components.js"></script>
+    <script src="/GXModules/Mollie/Mollie/Shop/Javascripts/mollie-credit-card.js"></script>
     <script src="https://js.mollie.com/v1/mollie.js"></script>
     <link rel="stylesheet" href="/GXModules/Mollie/Mollie/Shop/Styles/mollie-applepay.css">
